requirements.txt
accelerate
Flask
transformers
torch
flask-cors
gunicorn
numpy
scikit-learn
pandas
sentence-transformers
textblob
pyspellchecker